Split context summary boundary
