+++ Tests/QuillCodeParityTests/ParityWorkspaceModelThreadGateTests.swift
@@ -41,7 +41,11 @@ final class ParityWorkspaceModelThreadGateTests: QuillCodeParityTestCase {
         let creationText = try Self.appSourceText(named: "WorkspaceThreadCreationEngine.swift")
         let continuationText = try Self.appSourceText(named: "WorkspaceModelContextContinuations.swift")
         let threadExtensionText = try Self.appSourceText(named: "WorkspaceModelThreads.swift")
-        let summaryGeneratorText = try Self.appSourceText(named: "WorkspaceContextSummaryGenerator.swift")
+        let summaryGeneratorText = try Self.appSourceText(named: "WorkspaceContextSummaryGenerators.swift")
+        let summaryPromptText = try Self.appSourceText(named: "WorkspaceContextSummaryPromptBuilder.swift")
+        let summarySanitizerText = try Self.appSourceText(named: "WorkspaceContextSummarySanitizer.swift")
+        let summaryRequestText = try Self.appSourceText(named: "WorkspaceContextSummaryRequest.swift")
+        let summaryTelemetryText = try Self.appSourceText(named: "WorkspaceContextSummaryTelemetry.swift")
         let runtimeText = try Self.appSourceText(named: "RuntimeFactory.swift")
 
         XCTAssertTrue(seedBuilderText.contains("struct WorkspaceThreadSeedBuilder"), "Fork and compact seed construction should live in a focused builder.")
@@ -50,6 +54,10 @@ final class ParityWorkspaceModelThreadGateTests: QuillCodeParityTestCase {
         XCTAssertTrue(seedBuilderText.contains("static func compactSeedMessages"), "Compact seed construction should be directly testable.")
         XCTAssertTrue(summaryGeneratorText.contains("protocol WorkspaceContextSummaryGenerating"), "Model-backed context summaries should live behind a focused summary generator protocol.")
         XCTAssertTrue(summaryGeneratorText.contains("struct LLMWorkspaceContextSummaryGenerator"), "Live context summaries should be implemented outside the workspace model.")
+        XCTAssertTrue(summaryPromptText.contains("enum WorkspaceContextSummaryPromptBuilder"), "Context-summary prompt shaping should live outside generator orchestration.")
+        XCTAssertTrue(summarySanitizerText.contains("enum WorkspaceContextSummarySanitizer"), "Context-summary redaction and truncation should live outside generator orchestration.")
+        XCTAssertTrue(summaryRequestText.contains("struct WorkspaceContextSummaryRequest"), "Context-summary request DTOs should have a focused model file.")
+        XCTAssertTrue(summaryTelemetryText.contains("struct WorkspaceContextSummaryTelemetry"), "Context-summary telemetry DTOs should have a focused model file.")
         XCTAssertTrue(runtimeText.contains("contextSummaryGenerator: LLMWorkspaceContextSummaryGenerator"), "TrustedRouter runtime should wire model-backed context summaries.")
         XCTAssertTrue(creationText.contains("WorkspaceThreadSeedBuilder.forkSeedMessages"), "Thread creation should delegate fork seeding.")
         XCTAssertTrue(creationText.contains("WorkspaceThreadSeedBuilder.compactSeedMessages"), "Thread creation should delegate context compaction seeding.")
